move service_id to settings
