+++ cqrs-example-server/src/Settings.toml
@@ -3,5 +3,5 @@ response_topic = "TEST_RESPONSE"
 events_topic = "TEST_EVENTS"
 bootstrap_server = "127.0.0.1:9092"
 transaction_id = "transaction_id"
-service_id = "service_id"
+service_id = "CQRS-EXAMPLE-SERVER"
 log_level = "info"
