+++ cqrs-example-server/src/main.rs
@@ -3,16 +3,18 @@ mod prelude;
 
 use crate::prelude::*;
 use config::Config;
+use cqrs_kafka::KafkaSettings;
 use cqrs_kafka::inbound::StreamKafkaInboundChannel;
-use cqrs_kafka::outbound::{KafkaOutboundChannel, KafkaSettings, TransactionalKafkaOutboundChannel};
+use cqrs_kafka::outbound::TransactionalKafkaOutboundChannel;
 use cqrs_library::cqrs::command::{CommandAccessor, CommandStore};
 use cqrs_library::cqrs::messages::CommandResponse;
 use cqrs_library::cqrs::traits::{Command, Event, EventProducer};
-use cqrs_library::cqrs::{CommandServiceServer, CqrsEventProducer};
+use cqrs_library::cqrs::CommandServiceServer;
 use log::{error, info};
 use serde::{Deserialize, Serialize};
 use std::env;
 use std::sync::Arc;
+
 
 fn handle_create_user(
     command_accessor: &mut CommandAccessor,
@@ -109,7 +111,7 @@ async fn main() -> Result<()> {
         .create_topic(&kafka_settings.events_topic.as_str())
         .await?;
 
-    let mut command_store = CommandStore::new("COMMAND-SERVER");
+    let mut command_store = CommandStore::new(kafka_settings.service_id.as_str());
     command_store.register_handler("CreateUserCommand", handle_create_user);
 
 
@@ -121,7 +123,7 @@ async fn main() -> Result<()> {
     );
 
     let command_channel = StreamKafkaInboundChannel::new(
-        "COMMAND-SERVER",
+        kafka_settings.service_id.as_str(),
         &[kafka_settings.commands_topic.as_str()],
         &kafka_settings.bootstrap_server,
         Arc::new(command_service_server),
